refactored into generic decoupled module
