+++ modules/stanford_decoupled/src/Plugin/GraphQLCompose/EntityType/LayoutLibrary.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\stanford_decoupled\Plugin\GraphQLCompose\EntityType;
+
+use Drupal\graphql_compose\Plugin\GraphQLCompose\GraphQLComposeEntityTypeBase;
+
+/**
+ * {@inheritdoc}
+ *
+ * @GraphQLComposeEntityType(
+ *   id = "layout",
+ *   type_sdl = "LayoutLibrary"
+ * )
+ */
+class LayoutLibrary extends GraphQLComposeEntityTypeBase {
+
+}
