+++ modules/stanford_decoupled/src/Plugin/GraphQLCompose/EntityType/PolicyLog.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Drupal\stanford_decoupled\Plugin\GraphQLCompose\EntityType;
+
+use Drupal\graphql_compose\Plugin\GraphQLCompose\GraphQLComposeEntityTypeBase;
+
+/**
+ * {@inheritdoc}
+ *
+ * @GraphQLComposeEntityType(
+ *   id = "su_policy_log"
+ * )
+ */
+class PolicyLog extends GraphQLComposeEntityTypeBase {
+
+}
