--- modules/stanford_graphql/src/Plugin/GraphQLCompose/EntityType/Citation.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Drupal\stanford_graphql\Plugin\GraphQLCompose\EntityType;
-
-use Drupal\graphql_compose\Plugin\GraphQLCompose\GraphQLComposeEntityTypeBase;
-
-/**
- * {@inheritdoc}
- *
- * @GraphQLComposeEntityType(
- *   id = "citation"
- * )
- */
-class Citation extends GraphQLComposeEntityTypeBase {
-
-}
