--- modules/stanford_graphql/src/Plugin/GraphQLCompose/EntityType/ConfigPage.php
@@ -1,18 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Drupal\stanford_graphql\Plugin\GraphQLCompose\EntityType;
-
-use Drupal\graphql_compose\Plugin\GraphQLCompose\GraphQLComposeEntityTypeBase;
-
-/**
- * {@inheritdoc}
- *
- * @GraphQLComposeEntityType(
- *   id = "config_pages"
- * )
- */
-class ConfigPage extends GraphQLComposeEntityTypeBase {
-
-}
