--- modules/stanford_graphql/src/Plugin/GraphQLCompose/EntityType/LayoutLibrary.php
@@ -1,19 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Drupal\stanford_graphql\Plugin\GraphQLCompose\EntityType;
-
-use Drupal\graphql_compose\Plugin\GraphQLCompose\GraphQLComposeEntityTypeBase;
-
-/**
- * {@inheritdoc}
- *
- * @GraphQLComposeEntityType(
- *   id = "layout",
- *   type_sdl = "LayoutLibrary"
- * )
- */
-class LayoutLibrary extends GraphQLComposeEntityTypeBase {
-
-}
