--- modules/stanford_graphql/stanford_graphql.info.yml
@@ -1,10 +1,0 @@
-name: 'Stanford GraphQL'
-type: module
-description: 'GraphQL helper.'
-core_version_requirement: ^9 || ^10
-package: Stanford
-version: 9.4.1
-dependencies:
-  - graphql_compose:graphql_compose_layout_paragraphs
-  - graphql_compose:graphql_compose_menus
-  - graphql_compose:graphql_compose_views
