--- modules/stanford_graphql/stanford_graphql.module
@@ -1,59 +1,0 @@
-<?php
-
-/**
- * @file
- * stanford_graphql.module
- */
-
-use Drupal\graphql\GraphQL\Execution\FieldContext;
-use Drupal\graphql_compose\Plugin\GraphQLCompose\GraphQLComposeFieldTypeInterface;
-use Drupal\paragraphs\ParagraphInterface;
-
-/**
- * Implements hook_graphql_compose_field_results_alter().
- */
-function stanford_graphql_graphql_compose_field_results_alter(array &$results, $entity, GraphQLComposeFieldTypeInterface $plugin, FieldContext $context) {
-  $field_definition = $plugin->getFieldDefinition();
-  if ($field_definition->getName() == 'layout_selection') {
-    foreach ($results as &$result) {
-      $result = [
-        'id' => $results->id(),
-        'label' => $results->label(),
-      ];
-    }
-  }
-
-  foreach ($results as $result) {
-    if ($result instanceof ParagraphInterface) {
-      $behaviors = $result->getAllBehaviorSettings();
-      $result->set('behavior_settings', $behaviors ? json_encode($behaviors) : NULL);
-    }
-  }
-
-  $text_fields = ['text', 'text_long', 'text_with_summary'];
-  if (in_array($field_definition->getType(), $text_fields)) {
-    foreach ($results as &$result) {
-      foreach ($result as &$text) {
-        $text = preg_replace('/(\r\n)+|\r+|\n+|\t+/', ' ', (string) $text);
-        // Remove html comments.
-        $text = preg_replace('/<!--.*?>/', '', $text);
-        // Remove white space between tags.
-        $text = preg_replace('/> +?</', '><', $text);
-      }
-    }
-  }
-}
-
-/**
- * Implements hook_graphql_compose_entity_base_fields_alter().
- */
-function stanford_graphql_graphql_compose_entity_base_fields_alter(array &$fields, string $entity_type_id) {
-  if ($entity_type_id == 'paragraph') {
-    $fields['behavior_settings'] = [
-      'field_type' => 'string',
-      'name_sdl' => 'behaviors',
-      'required' => FALSE,
-      'description' => t('Paragraph Behavior Settings.'),
-    ];
-  }
-}
