+++ stanford_profile_helper.module
@@ -1088,42 +1088,6 @@ function stanford_profile_helper_entity_view(array &$build, EntityInterface $ent
 }
 
 /**
- * Implements hook_preprocess_HOOK().
- */
-function stanford_profile_helper_preprocess_image(&$variables) {
-  $decoupled = _stanford_profile_is_decoupled();
-
-  if ($decoupled && (empty($variables['width']) || empty($variables['height']))) {
-    $path = str_starts_with($variables['uri'], '/') ? DRUPAL_ROOT . $variables['uri'] : $variables['uri'];
-    $path = preg_replace('/\?.*$/', '', $path);
-    if ($size = @getimagesize($path)) {
-      $variables['attributes']['data-width'] = $size[0];
-      $variables['attributes']['data-height'] = $size[1];
-    }
-  }
-}
-
-/**
- * Implements hook_preprocess_HOOK().
- */
-function stanford_profile_helper_preprocess_file_link(&$variables) {
-  if (_stanford_profile_is_decoupled()) {
-    $variables['link']['#url']->setAbsolute();
-  }
-}
-
-/**
- * Check if the site is configured to be decoupled.
- *
- * @return bool
- *   True if it's intended to be decoupled.
- */
-function _stanford_profile_is_decoupled(): bool {
-  return (bool) \Drupal::state()
-    ->get('stanford_profile_helper.decoupled', FALSE);
-}
-
-/**
  * Implements hook_search_api_processor_info_alter().
  */
 function stanford_profile_helper_search_api_processor_info_alter(array &$processors) {
@@ -1135,6 +1099,8 @@ function stanford_profile_helper_search_api_processor_info_alter(array &$process
  */
 function stanford_profile_helper_entity_view_display_alter(EntityViewDisplayInterface $display, array $context) {
   if ($context['view_mode'] == 'search_indexing' && $context['entity_type'] == 'node') {
+    // The title is already in the template, it's not needed in the display.
     $display->removeComponent('title');
   }
 }
+
